add Spring to project
