+++ src/main/java/com/example/Develhope_Project/DevelhopeProjectApplication.java
@@ -1,0 +1,13 @@
+package com.example.Develhope_Project;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DevelhopeProjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DevelhopeProjectApplication.class, args);
+	}
+
+}
