+++ src/main/java/com/example/Develhope_Project/repository/Config.java
@@ -1,0 +1,15 @@
+package com.example.Develhope_Project.repository;
+
+public class Config {
+
+    public static String urlDB = "jdbc:mysql://localhost:3306/";
+    public static String NameDBdruwan = "team_project";
+    public static String UsernameDBdruwan = "root";
+    public static String PasswordDBdruwan = "Asdasd123stella.";
+    public static String NameDBcamu = "team_project";
+    public static String UsernameDBcamu = "root";
+    public static String PasswordDBcamu = "developerCamu*@";
+    public static String NameDBmike = "team_project";
+    public static String UsernameDBmike = "root";
+    public static String PasswordDBmike = "Maicoldevelhope12.";
+}
