+++ src/test/java/com/example/Develhope_Project/DevelhopeProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.Develhope_Project;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DevelhopeProjectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
